server/requirements.txt
asgiref==3.7.2
django==4.2.3
sqlparse==0.4.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
tensorflow==2.11.1
tensorflow-text==2.11.0
pandas==2.0.3
numpy==1.25.0